requirements.txt

openai
Flask
flask-cors
python-dotenv
